Clean up the sample
* Fixed code analysis warnings
* Use C++ to load script files
* Move helpers to separate header files
