+++ bin/Debugger.Sample/ChakraCore.Debugger.Sample.vcxproj
@@ -55,6 +55,9 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="DebugProtocolHandler.h" />
+    <ClInclude Include="DebugService.h" />
+    <ClInclude Include="ErrorHelpers.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
